requirements.txt
streamlit
pymongo
requests
datetime
Pillow